[MUI-414] Fixed options in AssetName story
